+++ tbot/machine/channel/subprocess.py
@@ -14,12 +14,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import typing
-import pty
-import os
 import fcntl
+import os
+import pty
 import select
 import subprocess
+import time
+import typing
+from tbot import log
 from . import channel
 
 
@@ -79,7 +81,7 @@ class SubprocessChannel(channel.Channel):
         try:
             buf = os.read(self.pty_master, maxread)
             self._debug_log(buf)
-        except BlockingIOError:
+        except (BlockingIOError, OSError):
             # If we don't get anything, and the timeout hasn't triggered
             # this channel is closed
             raise channel.ChannelClosedException()
@@ -100,8 +102,32 @@ class SubprocessChannel(channel.Channel):
     def close(self) -> None:  # noqa: D102
         if self.isopen():
             self.cleanup()
-        self.p.kill()
+        sid = os.getsid(self.p.pid)
+        self.p.terminate()
+        os.close(self.pty_master)
         self.p.wait()
+
+        # Wait for all processes in the session to end.  Most of the time
+        # this will return immediately, but in some cases (eg. a serial session
+        # with picocom) we have to wait a bit until we can continue.  To be as
+        # quick as possible this is implemented as exponential backoff and will
+        # give up after 1 second (1.27 to be exact) and emit a warning.
+        for t in range(7):
+            if (
+                subprocess.call(
+                    ["ps", "-s", str(sid)],
+                    stdout=subprocess.DEVNULL,
+                    stderr=subprocess.DEVNULL,
+                )
+                != 0
+            ):
+                break
+            time.sleep(2 ** t / 100)
+        else:
+            log.warning(
+                "There are still processes running in this session.\n"
+                "   Something might break."
+            )
 
     def fileno(self) -> int:  # noqa: D102
         return self.pty_master