+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - `selftest`s sometimes failing if dropbear does not start fast enough
 - `SSHMachine`s using the local user's home dir instead of the one on the
   lab-host.
+- Local channels now correctly end the session which fixes weird bugs like
+  picocom not being able to reaquire the shell.
 
 
 ## [0.7.0] - 2019-02-08